+++ src/Omni/Helper/BasketHelper.php
@@ -258,7 +258,10 @@ class BasketHelper extends AbstractHelper
 
         foreach ($quoteItems as $quoteItem) {
 
-            list($itemId, $variantId, $uom, $barCode) = $this->itemHelper->getComparisonValues($quoteItem);
+            list($itemId, $variantId, $uom, $barCode) = $this->itemHelper->getComparisonValues(
+                $quoteItem->getProductId(),
+                $quoteItem->getSku()
+            );
 
             // @codingStandardsIgnoreLine
             $list_item = (new Entity\OneListItem())
@@ -307,7 +310,10 @@ class BasketHelper extends AbstractHelper
             } else {
                 $product = $item->getProduct();
             }
-            list($itemId, $variantId, $uom, $barCode) = $this->itemHelper->getComparisonValues($product);
+            list($itemId, $variantId, $uom, $barCode) = $this->itemHelper->getComparisonValues(
+                $product->getId(),
+                $product->getSku()
+            );
             $qty     = $item->getData('qty');
             // @codingStandardsIgnoreLine
             $list_item = (new Entity\OneListItem())
@@ -464,7 +470,7 @@ class BasketHelper extends AbstractHelper
     }
 
     /**
-     * This function is overriding in hospitality module
+     * Send coupon code to basket calculation
      * @param $couponCode
      * @return Entity\OneListCalculateResponse|Phrase|string|null
      * @throws InvalidEnumException
@@ -498,16 +504,20 @@ class BasketHelper extends AbstractHelper
             $this->checkoutSession->getQuote()->getLsGiftCardNo(),
             $this->checkoutSession->getQuote()->getLsGiftCardAmountUsed(),
             $this->checkoutSession->getQuote()->getLsPointsSpent(),
-            $status
+            $status,
+            false
         );
 
-        if (!is_object($status) && $checkCouponAmount) {
+        if (!is_object($status) || $checkCouponAmount) {
             $this->couponCode = '';
             $this->update(
                 $this->get()
             );
             $this->setCouponQuote($this->couponCode);
             $status = __("Coupon Code is not valid");
+            if ($checkCouponAmount) {
+                $status = $checkCouponAmount;
+            }
             return $status;
         } elseif (!empty($status->getOrderDiscountLines()->getOrderDiscountLine())) {
             if (is_array($status->getOrderDiscountLines()->getOrderDiscountLine())) {
@@ -684,8 +694,8 @@ class BasketHelper extends AbstractHelper
      */
     public function getOneListAdmin($customerEmail, $websiteId, $isGuest)
     {
-        $list     = null;
-        $cardId   = null;
+        $list   = null;
+        $cardId = null;
 
         if (!$isGuest) {
             $customer = $this->customerFactory->create()->setWebsiteId($websiteId)->loadByEmail($customerEmail);
@@ -694,7 +704,7 @@ class BasketHelper extends AbstractHelper
                 $cardId = $customer->getData('lsr_cardid');
             }
         }
-        $webStore = $this->lsr->getWebsiteConfig(LSR::SC_SERVICE_STORE, $websiteId);
+        $webStore       = $this->lsr->getWebsiteConfig(LSR::SC_SERVICE_STORE, $websiteId);
         $this->store_id = $webStore;
         // @codingStandardsIgnoreStart
         /** @var Entity\OneList $list */
@@ -873,7 +883,10 @@ class BasketHelper extends AbstractHelper
     public function getItemRowTotal($item)
     {
         $baseUnitOfMeasure = $item->getProduct()->getData('uom');
-        list($itemId, $variantId, $uom) = $this->itemHelper->getComparisonValues($item);
+        list($itemId, $variantId, $uom) = $this->itemHelper->getComparisonValues(
+            $item->getProductId(),
+            $item->getSku()
+        );
         $rowTotal   = $item->getRowTotal();
         $basketData = $this->getOneListCalculation();
         $orderLines = $basketData ? $basketData->getOrderLines()->getOrderLine() : [];
@@ -918,13 +931,13 @@ class BasketHelper extends AbstractHelper
     public function calculateOneListFromOrder($order)
     {
         $couponCode = $order->getCouponCode();
-        $quote = $this->cartRepository->get($order->getQuoteId());
-        $oneList = $this->getOneListAdmin(
+        $quote      = $this->cartRepository->get($order->getQuoteId());
+        $oneList    = $this->getOneListAdmin(
             $order->getCustomerEmail(),
             $order->getStore()->getWebsiteId(),
             $order->getCustomerIsGuest()
         );
-        $oneList = $this->setOneListQuote($quote, $oneList);
+        $oneList    = $this->setOneListQuote($quote, $oneList);
         $this->setCouponCodeInAdmin($couponCode);
 
         return $this->update($oneList);