+++ src/Omni/Plugin/Checkout/Model/LayoutProcessorPlugin.php
@@ -3,6 +3,7 @@
 namespace Ls\Omni\Plugin\Checkout\Model;
 
 use \Ls\Core\Model\LSR;
+use \Ls\Omni\Helper\ContactHelper;
 use \Ls\Omni\Helper\Data;
 use \Ls\Omni\Helper\GiftCardHelper;
 use \Ls\Omni\Helper\LoyaltyHelper;
@@ -10,8 +11,7 @@ use Magento\Checkout\Block\Checkout\LayoutProcessor;
 use Magento\Framework\Exception\NoSuchEntityException;
 
 /**
- * Class LayoutProcessorPlugin
- * @package Ls\Omni\Plugin\Checkout\Model
+ * Interceptor to intercept LayoutProcessor on checkout to remove required sections
  */
 class LayoutProcessorPlugin
 {
@@ -34,25 +34,34 @@ class LayoutProcessorPlugin
     public $lsr;
 
     /**
-     * LayoutProcessorPlugin constructor.
+     * @var ContactHelper
+     */
+    public $contactHelper;
+
+    /**
      * @param Data $data
      * @param LoyaltyHelper $loyaltyHelper
      * @param GiftCardHelper $giftCardHelper
      * @param LSR $lsr
+     * @param ContactHelper $contactHelper
      */
     public function __construct(
         Data $data,
         LoyaltyHelper $loyaltyHelper,
         GiftCardHelper $giftCardHelper,
-        LSR $lsr
+        LSR $lsr,
+        ContactHelper $contactHelper
     ) {
-        $this->data = $data;
-        $this->loyaltyHelper = $loyaltyHelper;
+        $this->data           = $data;
+        $this->loyaltyHelper  = $loyaltyHelper;
         $this->giftCardHelper = $giftCardHelper;
-        $this->lsr = $lsr;
+        $this->lsr            = $lsr;
+        $this->contactHelper  = $contactHelper;
     }
 
     /**
+     * After plugin to intercept LayoutProcessor process method
+     *
      * @param LayoutProcessor $subject
      * @param array $jsLayout
      * @return array
@@ -62,22 +71,30 @@ class LayoutProcessorPlugin
         LayoutProcessor $subject,
         array $jsLayout
     ) {
+        $shippingStep = &$jsLayout['components']['checkout']['children']['steps']['children']['shipping-step'];
+        $billingStep = &$jsLayout['components']['checkout']['children']['steps']['children']['billing-step'];
+        $sideBar = &$jsLayout['components']['checkout']['children']['sidebar'];
+        $shippingAdditional = &$shippingStep['children']['shippingAddress']['children']['shippingAdditional'];
+
         if ($this->data->isCouponsEnabled('checkout') == '0') {
-            unset($jsLayout['components']['checkout']['children']['steps']['children']['billing-step']['children']['payment']['children']['afterMethods']['children']['discount']);
-        }
-        if ($this->loyaltyHelper->isLoyaltyPointsEnabled('checkout') == '0') {
-            unset($jsLayout['components']['checkout']['children']['steps']['children']['billing-step']['children']['payment']['children']['afterMethods']['children']['loyalty-points']);
-            unset($jsLayout['components']['checkout']['children']['sidebar']['children']['summary']['children']['totals']['children']['loyalty_points_label']);
-        }
-        if ($this->giftCardHelper->isGiftCardEnabled('checkout') == '0') {
-            unset($jsLayout['components']['checkout']['children']['steps']['children']['billing-step']['children']['payment']['children']['afterMethods']['children']['gift-card']);
+            unset($billingStep['children']['payment']['children']['afterMethods']['children']['discount']);
         }
 
-        if (isset($jsLayout['components']['checkout']['children']['steps']['children']['shipping-step']['children']['shippingAddress']['children']['shippingAdditional']['children'])) {
-            $jsLayout['components']['checkout']['children']['steps']['children']['shipping-step']['children']['shippingAddress']['children']
-            ['shippingAdditional']['children']['select_store'] = ['component' => 'Ls_Omni/js/view/checkout/shipping/select-store'];
+        if ($this->loyaltyHelper->isLoyaltyPointsEnabled('checkout') == '0' ||
+            empty($this->contactHelper->getCardIdFromCustomerSession())) {
+            unset($billingStep['children']['payment']['children']['afterMethods']['children']['loyalty-points']);
+            unset($sideBar['children']['summary']['children']['totals']['children']['loyalty_points_label']);
+        }
+
+        if ($this->giftCardHelper->isGiftCardEnabled('checkout') == '0') {
+            unset($billingStep['children']['payment']['children']['afterMethods']['children']['gift-card']);
+        }
+
+        if (isset($shippingAdditional['children'])) {
+            $shippingAdditional['children']['select_store'] =
+                ['component' => 'Ls_Omni/js/view/checkout/shipping/select-store'];
         } else {
-            $jsLayout['components']['checkout']['children']['steps']['children']['shipping-step']['children']['shippingAddress']['children']['shippingAdditional'] =
+            $shippingAdditional =
                 [
                     'component' => "uiComponent",
                     'displayArea' => 'shippingAdditional',
@@ -90,12 +107,15 @@ class LayoutProcessorPlugin
         }
 
         if (!$this->isValid()) {
-            unset($jsLayout['components']['checkout']['children']['steps']['children']['shipping-step']['children']['shippingAddress']['children']['shippingAdditional']['children']['select_store']);
+            unset($shippingAdditional['children']['select_store']);
         }
+
         return $jsLayout;
     }
 
     /**
+     * Check to see commerce services connection status
+     *
      * @return bool|null
      * @throws NoSuchEntityException
      */